Refactor error handling in CommandExecutor class
Previously, the code in CommandExecutor class returned the command output if its code was 0 or threw an error if the output was falsy. However, this didn't account for cases where the code was 0 but the stderr was present, which should also throw an error. This commit adds a check for stderr and throws an appropriate error message.